// src/shared/api/soil-lab/model/measurementListItemResponse.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * soil_laboratory
 * OpenAPI spec version: 0.1.0
 */
import type { MeasurementListItemResponseCreatedById } from './measurementListItemResponseCreatedById'
import type { MeasurementListItemResponseDeletedAt } from './measurementListItemResponseDeletedAt'
import type { MeasurementListItemResponseUpdatedById } from './measurementListItemResponseUpdatedById'

export interface MeasurementListItemResponse {
  deletedAt?: MeasurementListItemResponseDeletedAt
  createdAt: string
  createdById?: MeasurementListItemResponseCreatedById
  updatedAt: string
  updatedById?: MeasurementListItemResponseUpdatedById
  id: string
  moldingSandNumber: string
  moldingSandStrengthKgfCm2: number
  moldingSandGasPermeability: number
  moldingSandMoisturePercent: number
  readonly isDeleted: boolean
}